Decouple server creation and listening from Application
